requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.9
djangorestframework==3.4.7
idna==2.6
requests==2.18.4
urllib3==1.22
wheel==0.24.0
